refactor(ci): publish test results in a separate workflow
Publishing unit test results the way we did does not work when we get a PR from a forked repo (eg. dependabot). This is because forked repositories run with a read-only GITHUB_TOKEN.
We can circumvent this problem by publishing the test results in a separate workflow. This workflow gets triggered upon completion of the workflow that is triggered by the forked PR. The second workflow will have the right to write, thus is able to publish the test results. The first workflow has to upload its event file as an artifact for the second workflow to be able to publish the results.

For more info see: https://github.com/EnricoMi/publish-unit-test-result-action#support-fork-repositories-and-dependabot-branches