ci: run only 1 job in parallel
Due to the limited resources of the github runners executing the tests for all 3 scenarios could result into timeouts when sending message with the client. Changing this parallelisation to only run 1 job at a time will increase the workflow time to ~10 minutes, but should resolve the flakiness of the message tests.

In the future it would be nice to make client configuration customizable, so we could increase the request timeout for just the problematic tests, while still being able to run our jobs in parallel.