feat: move containerized engine to different module
A new "engine-agent" module has been created. This module will be responsible for containerizing the engine and making it executable. This has been split from the "engine" module, because in the near future we will support using the engine with and without testcontainers. For people running without testcontainers we don't want to expose any of the container related stuff.