feat: add server side grpc implementation
Adds the server-side implementation of the engine control. The engine control implementation lives in the engine-agent. The reason for this is that we only need to communicate with the engine through grpc when it is running in a testcontainer. The engine-agent is a wrapper around the engine that make it executable in docker and now is also responsible for the communication with the container.