ci: create filters module
A new module "filters" has been introduced. This module will be responsible for filtering the record stream to find specific records.
It has been extracted because this is a shared functionality, regardless of where the record stream comes from. In the future this might move to Zeebe as we have a similar functionality in that project which we could reuse.