feat: add extension using testcontainers
This extension makes it possible to run tests against a testcontainer. The extension is responsible for managing the testcontainer lifecycle. It will start the container once for the entire testsuite. This is done in the beforeAll() method. The EngineContainer singleton will make sure it only gets created once.

Before each test the engine gets reset. This is done in the beforeEach() method, instead of in the afterEach(). The reasoning behind this is that the jUnit lifecycle calls afterEach() before testFailed(). We want to access the record stream upon test failure. If we reset the engine in the afterEach() method we will lose our access to the record stream in the testFailed() method.