fix: map records in a threadsafe way
The toJson() method on Records is not threadsafe at the moment. This is because some Records contain an ArrayProperty. During serializationa new Record will be created (eg. a DeploymentRecord). This is an UnpackedObject which contains an ArrayProperty. The ArrayProperty has an ArrayValue.
Serializing an ArrayValue modifies the object. This means that multiple threads trying to serialize the same DeploymentRecord at a time will modify the ArrayValue at the same time. This results in exceptions.

Making it thread-safe this way a quick-fix. It would be nice if we can map and store the records as they come in. This way the requests themselves only have to return a list of mapped records, instead of mapping all records.