feat: duplicate tests and add testcontainers extension
We want to verify both our extensions. Therefore, it makes sense to run all our QA tests against both extensions. For now all the tests are duplicated to achieve this. This, of course, is not ideal in maintaining the tests. In https://github.com/camunda-cloud/zeebe-process-test/issues/195 we will get rid of this and find a more generic solution.