feat: get image and ports from properties file
The image name, tag and ports used were hardcoded. In our workflow we'll have to modify at least the tag so we can run QA tests against the correct versions. This is easier to do if we store these options in a properties file.

This commit also removed the BindableService vararg. This was never used.