ci: run qa tests using testcontainers
We have 2 kinds of tests in our QA module. The regular ones, and the ones that will run using testcontainers. The regular ones will be ran against 3 OS (ubuntu, windows and mac OS). This hard to achieve in a generic way with testcontainers. This is because with tetcontainers we should build the docker image first, so we ensure we run the tests against the latest code. There is multiple issues with this:

1. Windows uses the wrong daemon and is unable to build the image.
2. GitHub's macos-latest container does not have docker installed.

There is probably ways to switch the daemon and install docker. However, there is not much value in running all the tests 2 times on different OSs. Besides this, the execution will increase dramatically. Because of this we will run the testcontainer tests on 1 OS (ubuntu).