ci: depend maven release on docker release
We used to run the maven and docker release jobs simultaneously. Now that the docker job has been changed to update the configuration files we have to make sure the maven release job runs after this is finished. If we don't we introduce a race-condition.
If the maven release job completes before the docker job updated the configuration files, it results in the maven release running tests on the wrong container.