ci: fix sed commands
Variable substitution did not work properly. This had 2 reasons. First, the $IMAGE_ID and $TAG were part of the string and got replaced as these strings.
Second, the $IMAGE_ID contains a /. This is also the character used as a delimiter in sed. The delimiter has now been changed to :