feat: improved logging upon failure
It's important users are able to debug their process properly if a test fails. To enable this we should provide them with proper logging upon a failure.
In the past we used the CompactRecordLogger from zeebe-test-util for this. This solution had 2 problems:
1. zeebe-test-util is not bound to Java 8. This means users running on Java 8 cannot use it.
2. The CompactRecordLogger provides too much information for the targeted users of this project.

To solve these problems I have introduced a new logger. This will also log all records of the stream, but only logs the most important fields that users need for debugging their processes. At the moment it is a guess which fields these are. Once we receive user feedback we can easily modify the logging for the specific records that are not clear enough.