docs: add licensing folder
This folder contains the licenses that are relevant for the different parts of this project. Apache will be used for almost everything, as this project is not part of the core business for Camunda.
However, exceptions exist. Some of the modules are depending on the Zeebe engine. This is part of the core business of Camunda. Because of this, these modules are licens under the Zeebe license.