docs: add license information
In the README and CONTRIBUTING guide we should explain which parts of this project are licensed by which license.
The 3 exceptions are the modules that have dependencies on the Zeebe engine. Because the Zeebe engine is licensed under the Zeebe license, we cannot override this license here with the regular Apache license.