fix: prevent incident logging when resolved
We logged all incidents. This also included all incidents that were resolved. If the user resolved an incident then it's safe to assume they are aware of the it, and we don't need to log it anymore.
With this change we only log the incidents that have not yet been resolved.