fix: increase space between columns
Some record types, value types and intents did not fit in the dedicated space reserved for it.
The largest RecordType (COMMAND_REJECTION) is 17 characters. This fits in the 20 character range.
The largest ValueType (MESSAGE_START_EVENT_SUBSCRIPTION) is 32 characters. This fits in the 35 character range.
The larges Intent (CREATE_WITH_AWAITING_RESULT) is 27 characters. This fits in the 30 character range.