build(deps): remove slf4j implementation from compile scope
We shouldn't add a logging implementation to our dependencies. This is troublesome for users, as they'll have to the same logging implementation. Or they have to exclude the dependency manually.

By changing the compile time scoped dependencies to slf4j-api we can solve this problem. Wherever we need an implementation, we should add it to the test scoped dependencies.