build(deps): add zeebe dependencies
Starting from 1.4.0-alpha2 the zeebe release bom has been changed. Only the modules that are supposed to be used by users are now specified in there. This means that for all the modules that are being used here that have been removed from the bom, they need to be explicitly specified.

https://github.com/camunda-cloud/zeebe/pull/8833