ci: fix updating image tag in git when in headless mode
If the workfow is started by a new release it will checkout the tag associated with the release. This means git runs in headless mode and can't pull or push. We shouldn't try to commit the changes to the image id and tag when running in headless mode.

`git symbolic-ref -q HEAD` will return exit code 1 when in headless mode. By running this command in the if we can check the exit code. If it's 1 we do not need to commit and push.