ci: combine deploy to docker and deploy to maven jobs
In the "Update image and tag in Git" step we update the configuration file with the proper image name and tag. For workflows triggered by releases this part is done in HEADLESS mode. This means we cannot push these changes to a remote.

In the deploy-to-maven job we did a new checkout of the code. In this code the changes we made in te "Update image and tag in Git" are not available. This causes the application to be packaged whilst pointing to the wrong image.

By combining these 2 jobs into one we no longer do a 2nd checkout. As a result the application gets packaged with the correct configurations.