ci: filter extension-testcontainer resources
Adding this as a resources to maven and applying filtering allows maven to set the tag dynamically to the project.version.

When we create a new release we will build a new image tagged with the release version. Maven will build the application with the release version. By syncing these 2 up the version should always match.