ci: override tag to latest for workflow started by main
Maven overrides the image tag in the config with the project.version. This works fine for releases. Release X.Y.Z gets created, an image is built and pushed with tag X.Y.Z and maven updates the config to X.Y.Z and packages the application.

The main branch is an exception here. The project.version in this case would be X.Y.Z-SNAPSHOT. We don't want to publish a lot of different SNAPSHOT versions on docker hub. For this reason we should override it with latest. The workflow would then be as follows:

1. A change is pushed to the main branch
2. Everything gets tested and verified
3. A docker image is built and pushed with the tag "latest"
4. We update the configuration to set the tag to "latest"
5. Maven doesn't replace any of configuration and packages it using "latest".

If we don't add this step maven will replace the tag in the config to X.Y.Z-SNAPSHOT. This does not match the tag of the image we just built.