ci(fix): build engine-agent jar and dependencies
We cannot build the engine-agent without building some of it's dependencies first.
By not changing directory to engine-agent we can run this command in the root directory.
With -pl we specify that we only want to build a specified project.
:zeebe-process-test-engine-agent references the artifact id of this project
With -am we specify that we also want to build the other projects that are required by the project specified with -pl