fix: add slf4j-simple dependency to engine-agent
This dependency had been changed to slf4j-api to prevent bothering users with an implementation of slf4j. However, the engine-agent is not supposed to be included as a dependency by users. It's sole purpose is running the engine in a testcontainer. Doing this requires an implementation of slf4j.