chore: add NOTICE file
Add a NOTICE file as is required by the Apache license, and because of this als by the Zeebe Community License.

In this file we credit other authors for code that we have used. I thought about crediting EZE and Camunda Bpmn Assert in this file. However, this is a bit strange considering both these projects are also owned by Camunda Services GmbH.