docs: add testcontainers dependency
Documents how to use the testcontainers version of ZPT. I've named the differences between the 2 dependencies and pros/cons for using both of them.
I've also provided some context on how to switch between the both versions in order to guide users into making their decision on which one to use.