refactor: add abstract test classes
Abstract test classes should contain all the test implementations. The other classes that extend this only need to provide the extension. This way we don't have to duplicate all of our test implementations.

One thing to note is that injection of the ZeebeClient, ZeebeTestEngine and RecordStream. These fields get injected in what jUnit defines as the test class:
final Class<?> requiredTestClass = extensionContext.getRequiredTestClass();

We are dealing with 2 different scenarios. We have the assertion tests. Here the abstract classes contain 2 nested classes, HappyPathTests and UnhappyPathTests. In this scenario the nested class is the test class. These fields get injected just fine.
In other tests we do not have any nested classes. Here the class that extends the abstract class will be the test class (e.g. the WorkterTest class). For these classes the fields must be declared on the extending class. We can access these in the abstract class by adding abstract getters.