ci: update versions on stable branch
The maven action plugin used to only be able to update the version on the default branch of the repository. Recently a change has been made to make the branch configurable. For this reason a step has been introduced to the workflow to decide our git branch.

It works by first figuring out what the major and minor version of the release are (e.g. 1.4 when the release is 1.4.1). If this fails to yield a result we exit the workflow. Otherwise we will add the branch (stable/major.minor) as an output of this step. This will then be used by the community-action-maven-release as the branch for updating the versions.