fix: set branch when running on main
We would only decide the git branch when the workflow was not triggered by a new release. When it was triggered by a push to the main branch it would not be set.

The community-action-maven-release does require the branch to contain a value. This means we need to set it for the main branch to the default branch.