feat: implement ColumnFamily API changes
The ColumnFamily interface in Zeebe has been modified. Before it only contained a put method. Starting from version 8.0.0 this has been replaced with update, insert and upsert methods.
The delete method has been replaced with deleteExisting and deleteIfExists.

In our in memory database we implement this interface. When updating to Zeebe 8.0.0 these changed need to be merged to keep the code from breaking.