chore: update Zeebe version
The api changes to ColumnFamily are first available in Zeebe 8.0.0-rc1. We need to use this version to create our own release candidate.
Upon upgrading I found there was another breaking change made in zeebe-protocol-jackson. For this reason a small change has been made to the way we do the mapping of records in ContainerizedEngine.