feat: extract called process instance
When a process contains a call activity it will call another process instance. If we want to assert this process we need to create a new ProcessInstanceAssert object, with the process instance key of the called process instance. Thus far we had no way to do this, except for using inspections. This was a bit of a hassle. With this new method we can now assert that a process instance has called a different process and using extractingLatestCalledProcess we can obtain the assertions from this process. This fits the other extracting... methods we have for asserting over e.g. incidents.