fix: prevent completion of already completed jobs
Our utility method completeTask would look for any jobs with a specific elementId that were created. It would not filter out jobs that were already completed. Because of this, if a process passed an element multiple times, we were only able to complete one job.

This fix filters out the jobs that were already completed. This way we can keep completing the task until there are no more jobs available.