build(.idea): restrict google-java-format version
The version used as the google-java-format plugin must be the same as
the one used in maven, as newer versions may lead to different
formatting.

Since we are currently using version 1.12.0, we should align the
IntelliJ plugin to the same version.

We don't want to restrict a maximum version, because some users might
already have updated to the latest version. That should be fine.
Especially, since we expect to be using the latest version soon as well.

Note that the IntelliJ plugin has an additional `.0` postfix to the
version number.