test: all gateway endpoints are implemented
This test checks if all the gateway endpoints are implemented. Since gRPC does not provide an interface which can force this during compile time we are at risk that updating the Zeebe version could result in ZPT not supporting all endpoints that can be triggered from the client. To circumvent this problem this test was created to make sure we implement all the methods gRPC expects us to.

(cherry picked from commit 3b76c460cd62523ba1ff7e8a511fec7365e6d25a)