refactor: move rpc response mapping
The response mapping has been moved from the GrpcResponseWriter to the GrpcToLogStreamGateway. The writer still contains the mapping functions, but they are called from the gateway.

This change is necessary in order to support both the DeployProcessCommand as the DeployResourceCommand. In the writer we don't have any context about the request available. Therefore it is unknown which request we received, and also which response we should return. The mapping was based on the ValueType of the record which is the same for both of these commands.

By letting the gateway decide which mapper should be used this problem has been resolved. At the point of entry of the request we will also decide which mapper should be used for the response. We nog longer need to decide the mapper base on the ValueType.

(cherry picked from commit 5a8d1a7261b8a8ede5f2ce049b55705b70b0f0f0)