test: add tests for RecordStreamLogger
This class was mostly not tested. We tested if we had a logger for all records, but we did not test the fucntionality of the logger itself. It doesn't contain much logic but by adding some tests around it we can hopefully prevent NPE's in the future.

(cherry picked from commit ba7227d30751dc4c3c2dfcb564ba98c5fe8ff641)