feat: don't start engine upon starting testcontainer
By not immediately starting the engine upon starting the container we become more flexible with the lifecycle of the engine in our extensions. This allows us to start it when we want to, instead of being tied to the testcontainer which has a different lifecycle from the engine.

(cherry picked from commit 8ea4c0db13d2fddc1be01b2b13cff4ba473c49fc)