feat: change testcontainer engine lifecycle
The start of the engine was tied to the start of testcontainer. This resulted in the following lifecycle:
1. BeforeAll - start the testcontainer and the engine
2. BeforeEach - stop the engine, recreate it and start it (reset)

This causes some strange behaviour. Upon test start we are starting the engine and resetting it without using it. With this change the lifecycle becomes somewhat more natural:
1. BeforeAll - start the testcontainer
2. BeforeEach - start the engine
3. AfterEach - stop the engine and recreate it (new reset)

This new lifecycle also matches more closely with the embedded lifecycle where we do:
1. BeforeEach - create and start the engine
2. AfterEach - stop the engine

(cherry picked from commit 687bf5bd210a670cc8e6c87c27e6c3bf505735c5)