build: don't provide test scope in dependencyManagement
Doing this could result in issues. For example:

The zeebe-workflow-engine has commons-lang3 as a dependency. Since it is in our dependencyManagement with a test scope, this means zeebe-worklfow-engine will get this dependency test scoped as well. This results in ClassNotFoundExceptions, since zeebe-workflow-engine expects it to be available on the classpath when running the application regularly.