feat: add storage class for requests to gateway
The GatewayRequestStore is a utility class that will be used by the GrpcToLogStreamGateway and the GrpcResponseWriter to share request between the two of them. The GrpcToLogStreamGateway will store new requests in here. The GrpcResponseWriter will take these requests out and write the correct response to the correct request.

This store has to be thread-safe. Because of this we are using a ConcurrentHashMap as well as an AtomicLong

(cherry picked from commit 95bd731f459dfe3bd6be0be64d2f1905aa8a86a2)