fix: remove single threaded executor
The single threaded executor was causing issues in combination with the GrpcResponseWriter. The response was send in a different thread from where the request was received. Since the writer reuses a DirectBuffer to store the record this could result in the buffer being modified by a thread, whilst the response was still being send in the single executor thread.

(cherry picked from commit e7826354dc3c6f5f46ea89728cbcdc193adfc2ca)