refactor: extract mapping of request responses to separate class
The GrpcResponseWriter used to do all the mapping of the responses. This should not be the responsibility of this class. It should only be responsible for writing the responses to the correct requests.

The mapping itself has now been extracted into a separate GrpcResponseMapper class.

(cherry picked from commit a6ff9d603f77b5918034f5e67063e6933c6479c4)