refactor(filters): support logRecord tests
The logRecords method could never be used to test because it only has
side effects (printing to log), but internally it works by building up a
String with a StringBuilder. By extracting this string building into a
separate method, its easy to test the produced String that will be
logged.