fix(ci): fix version decision command
The command to increase the Zeebe minor by 1 and appending -SNAPSHOT was broken. echo $(echo $CURRENT_ZEEBE_VERSION | awk -F. '/[0-9]+\./{$2++;print}' OFS=.)-SNAPSHOT would result in bash trying to execute the version. This would result in the message that 8.2.0-SNAPSHOT is not a command.

By moving the $() outward and using backticks this problem is resolved. $(echo `echo $CURRENT_ZEEBE_VERSION | awk -F. '/[0-9]+\./{$2++;print}' OFS=.`-SNAPSHOT) will echo the version instead of trying to execute it.