ci: add build-test job for bors to await
Bors needs a GH job that it can await and check the result of. Since
there is a matrix of testing related jobs in the build-test.yml
workflow, we need a final job that is only executed once all the other
test jobs have completed.

We also use this approach in the zeebe repo.