ci: add bors to merge our pull requests
Bors requires a bit of configuration. This configuration is mostly
copied from the Zeebe repo.

The status list specifies the CI jobs that bors will await. This result
determines whether or not bors will actually fast-forward the target
branch. The entry "Test summary" is the name of a job added to the
build-test.yml workflow. A comment is placed in that job definition to
highlight that bors depends on it.

The other properties are pretty straightforward.

(cherry picked from commit 2ff389d9e28ff9d5d44278a18adc297a24981e19)