fix(deps): fix dmn model dependency convergence issue
The Zeebe version is set to the SNAPSHOT version in order to resolve this issue. If we don't the main branch would expect version 7.17.0 of the dmn-model. The nightly tests against SNAPSHOT would expect version 7.18.0. Whichever version we'd pick, one of those would always be broken. By changing te main branch to Zeebe SNAPSHOT we resolve this issue.