fix(ci): remove caching from the Zeebe SNAPSHOT tests
The Zeebe SNAPSHOT tests are supposed to verify that zeebe-process-test is compatibly with the latest changes in Zeebe. It runs daily to make sure nothing has broken.
Caching messes this up. If we restore the cache we are no longer testing against the latest changes. We might be testing against an older cache, missing any changes that could break this project.