ci: use major version for actions/setup-java
It might be better to switch the version to the latest major release v3.
Such tags are generally moved with every new minor and patch release of
the major version. This practice is officially recommended by GH.

Note that this may result in reduced reproducibility. However, since all
tags can be moved, the only way to guarantee reproducibility is by
pinning specific commit shas.

Likely it is easiest for us when using actions like these (e.g.
checkout) to use the major version tag.