fix(extensions): inject ZPT fields into superclasses
The extensions should inject the engine, client and recordstream into the fields that are available. Before this only tried to inject them into the annotated class. This does not work when inheritance is used and the fields are defined on the superclass of the annotated class.

This commit changes the extension so that it will first look on the annotated class. If it cannot find any of the annotations here, it will work it's way up to the superclass and so on until it either found a field it is able to inject, or no fields at all.

(cherry picked from commit b912ba9702d81742f7a291d9e18d1db2b023be48)