feat: allow iteration to start at key
This implements the whileTrue(startAtKey, visitor) method, by
introducing a new whileEqualPrefix method that allows setting a startAt
key.

For the whileEqualPrefix method that does not take a startAt key, we can
simply use the prefix key as the starting position. See line 221. This
is similar to the implementation in Zeebe (see ref 1).

To enable iterating starting at a specific key, we need to change the
initial seek to use the startAt key instead of the prefix key. See line
240. This is inline with the implementation in Zeebe (see ref 2).

The startAt key is just an arbitrary key. In order to use it in the
seek, we need to prepend the logical column family id to the startAt
key. See lines 237-238 and the method `keyWithColumnFamily` in
`InMemoryDbColumnFamilyIterationContext`. This is inline with the
implementation in Zeebe (see ref 3). For a more in-depth explanation
please have a look at ref 4.

References:
1. https://github.com/camunda/zeebe/pull/11785/commits/f227c3201ef14764dd2c000e5cf2d229482f76d6#diff-d12a85834691a282f898186d31bd219e4e96e0f996fe98a6f7faa4cb56fb088dR324-R325
2. https://github.com/camunda/zeebe/pull/11785/commits/f227c3201ef14764dd2c000e5cf2d229482f76d6#diff-d12a85834691a282f898186d31bd219e4e96e0f996fe98a6f7faa4cb56fb088dL340-R355
3. https://github.com/camunda/zeebe/pull/11785/commits/f227c3201ef14764dd2c000e5cf2d229482f76d6#diff-06cd30102e4af6da91ba5a3c3d43c65b2e257557da286d242b9758588c30cc43R118-R125
4. https://github.com/camunda/zeebe/pull/11785#discussion_r1119000103