fix: implement whileTrue starting at key
This is a new method. To make the compilation pass, we need to implement
it. However, this is not yet possible because we lack the ability to
iterate starting from a specific key. Currently, all iterations start at
the beginning of the column family. We need to add this ability in the
upcoming commits.