fix: log command_distribution record value
The COMMAND_DISTRIBUTION value type was recently added. Which needs to
be covered by the record value loggers. This adds that logger.

Note, that I've chosen to only log 2 out of 3 properties:
- partition id (logged)
- value type (logged)
- command value (not logged)

We could choose to log the command value as part of the command
distribution records, but this could contain any record value. The
implementation would be more complex, as we cannot directly re-use the
registered valueTypeLoggers due to these methods requiring a record (not
a record value). Additionally, we don't know what this would look like
for random distributed commands.