refactor: move ObjectMapperConfig from api to assertions module
Into api module should be placed only classes that are injected into
user's test classes. Place the congifuration under the assertions
module, as it is can be accessed by VariableMapAssert

Change the stored object in the ObjectMapperConfig from ObjectMapper to ZeebeObjectMapper.
With this refactor we can remove the static code block inside VariablesMapAssert