feat: make ObjectMapper config thread-safe
The ObjectMapper is a static field. We can't guarantee that all test cases use the same ObjectMapper. If we don't store the mapper in a ThreadLocal, it could cause other testcases to interfere with a tests ObjectMapper, overriding it with a different value.

By using a ThreadLocal all testcases (running in their own thread) have their own ObjectMapper and other tests won't be interfering with this.

I've also done some slight refactorings to the extensions.