refactor: simplify adding other ignored methods
Currently, only the bindService method needs to be ignored as it is
explicitly declared by GatewayImplBase.

But the other methods from the interfaces BindableService and
AsyncService are not retrieved by getDeclaredMethods. For this, we'll
need to use getMethods. However, that method will also pass undeclared
methods like equals which we'll need to explicitly ignore.

This we can easily achieve using a list of strings.

(cherry picked from commit b75727036c14ec8b97871100b0a8cdd1768f07e7)