test: ensure all gateway rpcs supported
This test was not testing what was intended. It wants to ensure that all
the RPCs of the gateway are supported by the GrpcToLogStreamGateway. It
does this by verifying that the GrpcToLogStreamGateway declares an
implementation for each of the gateway's methods.

However, the getDeclaredMethods doesn't return those. Possibly these
were moved to AsyncService at some point. Note that this is generated
code and that this may have been changed by changes upstream in the grpc
library.

Instead, we must use getMethods to retrieve all methods available on
this class (also those defined through default implementations on its
interfaces). When we do that, we must also ignore a bunch of undeclared
methods, like equals and hashCode.

What we're left with is a failing test that fails for two missing
methods:
- streamActivatedJobs
- migrateProcessInstance

(cherry picked from commit 00b726b5d6be8621d6c48b3eb683bc39bced722d)