fix: ignore unsupported streamActivatedJobs RPC
It's not yet possible to support this method as it functions differently
from the other RPCS.

Typically, the RPCs are implemented by appending a command to the log
stream, but for streaming RPCs it works differently. Instead, the
gateway opens a stream through which the broker can stream events. As I
understand the implementation in Zeebe, this works without appending a
command.

In order to support this in ZPT, some new mechanism must be added that
allows streaming events to the calling client.