fix: unflake test by setting TTL to 0
ZPT test cases that also use timetravel should either:
- not buffer messages
- buffer messages for longer then they travel through time

When a buffered message expires during time travel, this can be detected
by ZPT as busy and idle state, rather than the actual event your test is
waiting for like a timer event that should trigger.

This can be the cause of flaky tests, and was so for us, because we run
these example test cases as well.