test: unflake test case
The exported stream may have a short delay before it sees records. This
means that the test should wait for some time between joining on the
client command to migrate, and asserting that some record is exported.

This test case should not verify the exported stream, because that's not
what these tests check. The test should focus on the usage of the
client and the response it sends rather than on the results it produces
in the engine. We have other tests that verify correct behavior inside
Zeebe.