fix(engine): Set variables from ThrowError command
Adjust the handling of the ThrowError gRPC command to set the variables from the command to the job record. As a result, the variables are set as local variables of the error catch event.

(cherry picked from commit bc7f8d887a23c214a2e4bb53b09096dd20a8832c)