fix: quick fix for a timer bug
Fixed rate timers don't play nicely with backwards
time adjustments. For fixed rate timers, an execution
is scheduled for ex. 5 seconds in the future. When that
execution runs, it schedules the next one, etc. However,
when traveling backwards in time, the next scheduled
execution never triggers.

The issue needs to be fixed properly in Zeebe. A temporary
workaround is to adjust the tests with a future date time.