fix: don't pass actor scheduler to logstream
Logstream no longer requires the actor scheduler, and this method
`.withActorSchedulingService` has been removed with a recent PR (see
below).

We can safely remove this from our logstream builder in the same way as
this was done for production zeebe, as the logstream no longer depends
on scheduling tasks in the future. Rather it has become completely
synchronous. This should be transparent to ZPT, but wasn't as the
logstream builder required passing the actor scheduler until now.

See: https://github.com/camunda/camunda/pull/22870