build: run TC tests with Java 21
We specify the release version to be Java 8 when running the tests,
which ensures that we will catch issues where use APIs > JDK 8, or
syntax that is unsupported on JDK 8, while still allowing us to use a
build environment (including plugins) for greater versions.

The one downside is we won't detect anymore if we pull in transitive
dependencies > JDK 8, but this should be detected separately.

(cherry picked from commit 0e8632fa52fd5fb6d1ae6c1d400a268969ef594c)