fix: log all unknown records using toString
Every time a record value is added, the RecordStreamLoggerTest fails
because we need to implement a nice record logger for that record value.

However, every time we just implement it with toString, which internally
uses toJson to fix the failing test.

Instead, of failing every time, we can just accept that new records are
logged as json.

This commit does this by changing the default logger to
Objects::toString (which uses toJson in the internal implementation of
our records).

We can also drop the valueTypeLogger for SBE_UNKNOWN as the unknown case
is now covered more generally.

This allows us to use this value type in a new unknown record that we
log and verify the logged output for.